requirements.txt
flask==2.3.3
gunicorn==23.0.0
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.0
scikit-learn==1.3.2